demo/backend/requirements.txt
flask
flask-cors
torch
torchvision
opencv-python
pillow
numpy
python-dotenv
